Implements IP address look up and handshake.
Host enumerates network interfaces, listens on all valid IPs.
Client enumerates network interfaces and sends messages from each
to the user-specified host IP.

When the host receives a valid message
it stores the socket that received the message as the socket to listen
on. All other sockets are closed and disposed of. The host then sends
a message back to the address that sent the message. When the client
receives this reply it makes not of the socket that received the reply
and closes and disposes off all other sockets. At this point we have a
bi-directional UDP socket connection between two computers.