Adds basic disconnect detection.
On the server side, once a connection is detected we start an timer
that tracks how long since the last message from a client. Once this
time exceeds a time out threshold we kill the connection.

On the client side once we start to attempt a connection we start a
timer that tracks how long since the last message from the host. This
allows us to gracefully timeout if we cannot establish the initial
connection. Otherwise timeouts work like the server.

Also updates the in-app log to tail.